Refactor H2Example to leverage Example test interface.
